+++ CHANGELOG.md
@@ -23,6 +23,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/virtual-users/branches/compare/master%0Drelease-3.3.6
 
+### Fixed
+- Install ChromeDriver only once instead of for every virtual user. Fix [JPERF-351].
+
+[JPERF-351]: https://ecosystem.atlassian.net/browse/JPERF-351
+
 ## [3.3.6] - 2019-01-04
 [3.3.6]: https://bitbucket.org/atlassian/virtual-users/branches/compare/release-3.3.6%0Drelease-3.3.5